+++ Docs/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Change log
 
-### v11.2.9.1001 (Interim release for 11.3.0.0)
+### v11.3.0.1001-beta Predictive Intellisense MetaDataFactory skeleton classes created.
 __New/ Enhancements__</br>
+-- Improved RunningConnectionTable drift detection.</br>
+-- Created Firebird Intellisense MetaDataFactory skeleton classes for predictive text.</br>
 -- Implemented Phase 1 of Firebird Language Service.</br>
 -- Added a user option for selecting the mandated language service for the`SqlEditor`. The available language services are `Firebird-SQL`, `Transact-SQL SSDT`, `Transact-SQL90` and `Unified-SQL Azure Data Lake`.</br>
 -- Added language service preferences for `Firebird-SQL`.</br>